Move thread_name to boost/capy/test
